Split out setax
